Move TokendClient related behavior tests into their own test class.
